+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0-beta.1
+
+- added argument parsing and help menu
+- implemented part of apt manager
+- implemented --version flag
+- added deduplication (may add logging at some point, will require different method)
+
 ## 2.0.0-beta.0
 
 - initial version